### 2016-09-08 #58 @whedon
[REVIEW]: SEP: Source Extractor as a library
**Submitting author:** @kbarbary (<a href="http://orcid.org/0000-0002-2532-3696">Kyle Barbary</a>)
**Repository:** <a href="https://github.com/kbarbary/sep" target ="_blank">https://github.com/kbarbary/sep</a>
**Version:** v1.0.0
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.159035" target ="_blank">http://dx.doi.org/10.5281/zenodo.159035</a>

**Reviewers and authors**: Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue
in the <a href="https://github.com/kbarbary/sep" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers)
in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice
versa will create corresponding breadcrumb trails in the link target.)
