### 2016-09-08 #58 @danielskatz
I feel like the original reference
Bertin, E. & Arnouts, S. 1996: SExtractor: Software for source extraction, Astronomy & Astrophysics Supplement 317, 393.  (doi: 10.1051/aas:1996164)
perhaps should be included, but this is up to the submitter.
