### 2016-09-08 215f7c2
paper: add citation to Bertin (1996)
