+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@ARTICLE{bertin96a,
+@ARTICLE{bertin96,
    author = {{Bertin}, E. and {Arnouts}, S.},
    title = "{SExtractor: Software for source extraction.}",
    journal = {\aaps},
@@ -7,6 +7,7 @@
    month = jun,
    volume = 117,
    pages = {393-404},
+   doi = {10.1051/aas:1996164},
    adsurl = {http://adsabs.harvard.edu/abs/1996A\%26AS..117..393B},
    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
