+++ README.md
@@ -98,7 +98,7 @@ Citation
 
 If you use SEP in a publication, please cite the following DOI. The link provides a variety of citation styles and BibTeX export.
 
-[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.15669.svg)](http://dx.doi.org/10.5281/zenodo.15669)
+[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.61658.svg)](http://dx.doi.org/10.5281/zenodo.61658)
 
 You may also wish to cite the original Source Extractor paper (Bertin 1996).
 
