### 2016-09-08 #58 @kbarbary
Thanks for the review!

**License:** I removed the copy of the BSD license and corrected the file extensions of the MIT license to `txt`.

**versions:** I updated the link in the readme to point to the v0.6.0 zenodo archive. (v1.0.0 will be the next release of SEP which will remove deprecated features; it has not yet been released though).

**citation:** I added the Bertin & Arnouts (1996) citation to the paper.
