### 2016-09-08 #58 @kbarbary
Ah, now I remember why the BSD license was there. One file `src/overlap.h` is based on BSD-licensed code. I believe this means I should keep the text of the BSD license there.

I could change MIT -> BSD to simply things overall; I just had a slight preference for MIT, so I started with that.
