### 2016-09-08 #58 @danielskatz
so the readme should be changed to say this (not the details, but just that there is some BSD code too)
