### 2016-09-09 f4f2800
clarify reason for BSD license in README
