### 2016-09-09 #58 @arfon
Do we need a updated paper compiling?
