### 2016-09-09 #58 @arfon
@danielskatz here you go: [10.21105.joss.00058.pdf](https://github.com/openjournals/joss-reviews/files/462979/10.21105.joss.00058.pdf)
