### 2016-09-10 #58 @danielskatz
re a statement of need in the paper: the repo readme probably should have a little of the paper content to explain what source extractor is and who might want to use it.  This could also potentially go in docs/index.srt

The license info in docs/index.srt might need to be updated?

It would also be useful if the docs contained a simple tutorial/test - just one input, the code needed to do an analysis, and a correct output to compare against.
