### 2016-09-10 #58 @danielskatz
re Community guidelines - the answers to the following are currently no.

Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Perhaps some more can be added to the README?  or to a CONTRIBUTING file?
