### 2016-09-10 #58 @jkahn
@danielskatz this is really useful feedback, but [in my role as self-appointed derail police](https://github.com/openjournals/joss/issues/163) please consider filing a ticket against [the target repo](https://github.com/kbarbary/sep) bug tracker and linking to this issue instead of pasting the body of error messages into this thread.
