### 2016-09-10 #58 @danielskatz
responding in the other issue (#163)
