### 2016-09-22 5bd678d
README improvements
- expanded "about" section
- more complete install instructions
- "contributing" section
