+++ README.md
@@ -7,25 +7,43 @@ Python and C library for Source Extraction and Photometry
 [![Build status](https://img.shields.io/appveyor/ci/kbarbary/sep.svg?style=flat-square&label=windows)](https://ci.appveyor.com/project/kbarbary/sep/branch/master)
 [![PyPI](https://img.shields.io/pypi/v/sep.svg?style=flat-square)](https://pypi.python.org/pypi/sep)
 
-
 *"... [it's] an SEP: Somebody Else's Problem."  
 "Oh, good. I can relax then."*
 
-[Source Extractor](http://www.astromatic.net/software/sextractor) is
-great, but sometimes you want to use a few of the pieces from it
-without running the entire executable. SEP makes available some of the
-algorithms in SExtractor as stand-alone functions and classes. These
-operate directly on in-memory arrays (no FITS files, configuration
-files, etc). The code is derived directly from the Source Extractor
-code base.
 
-SEP can be used from either Python or directly from C. See below for
-language-specific build and usage instructions.
+About
+-----
+
+[Source Extractor](http://www.astromatic.net/software/sextractor)
+(Bertin & Arnouts 1996) is a widely used
+command-line program for segmentation and analysis of astronomical
+images. It reads in FITS format files, performs a configurable series
+of tasks, including background estimation, source detection,
+deblending and a wide array of source measurements, and finally
+outputs a FITS format catalog file.
+
+While Source Extractor is highly useful, the fact that it can only be
+used as an executable can limit its applicability or lead to awkward
+workflows. There is often a desire to have programmatic access to
+perform one or more of the above tasks on in-memory images as part of
+a larger custom analysis.
+
+**SEP makes the core algorithms of Source Extractor available as a
+library of stand-alone functions and classes.** These operate directly
+on in-memory arrays (no FITS files or configuration files).  The code
+is derived from the Source Extractor code base (written in C) and aims
+to produce results compatible with Source Extractor whenever possible.
+SEP consists of a C library with no dependencies outside the standard
+library, and a Python module that wraps the C library in a Pythonic
+API. The Python wrapper operates on NumPy arrays with NumPy as its
+only dependency. See below for language-specfic build and usage
+instructions.
+
 
 Python
 ------
 
-**Documentation:** http://sep.readthedocs.org/
+**Documentation:** http://sep.readthedocs.io
 
 **Requirements:**
 
@@ -34,9 +52,25 @@ Python
 
 **Install release version:**
 
+SEP can be installed with [pip](https://pip.pypa.io). After ensuring
+that numpy is installed, run
+
 ```
-pip install sep
+pip install --no-deps sep
 ```
+
+If you get an error about permissions, you are probably using your
+system Python. In this case, I recommend using [pip's "user
+install"](https://pip.pypa.io/en/latest/user_guide/#user-installs)
+option to install sep into your user directory:
+
+```
+pip install --no-deps --user sep
+```
+
+Do **not** install sep or other third-party Python packages using
+`sudo` unless you are fully aware of the risks.
+
 
 **Install development version:**
 
@@ -84,6 +118,31 @@ and header file in `/path/to/prefix/include`. The default prefix is
 **API:** The C library API is documented in the header file
 [sep.h](src/sep.h).
 
+
+Contributing
+------------
+
+Report a bug or documentation issue: http://github.com/kbarbary/sep/issues
+
+Development of SEP takes place on GitHub at
+http://github.com/kbarbary/sep.  Contributions of bug fixes,
+documentation improvements and minor feature additions are welcome via
+GitHub pull requests. For major features, it is best to open an issue
[…]
