### 2016-09-22 9f9faf9
update docs (about, install, contributing) based on README
