+++ docs/index.rst
@@ -1,12 +1,36 @@
 SEP
 ===
 
-*Source Extraction and Photometry in Python*
+*Python library for Source Extraction and Photometry*
 
-SEP makes available some of the algorithms in Source Extractor as
-stand-alone functions and classes. These operate directly on in-memory
-numpy arrays (no FITS files, configuration files, etc). It's derived
-directly from (and tested against) the Source Extractor code base.
+
+About
+-----
+
+`Source Extractor <http://www.astromatic.net/software/sextractor>`_
+(Bertin & Arnouts 1996) is a widely used
+command-line program for segmentation and analysis of astronomical
+images. It reads in FITS format files, performs a configurable series
+of tasks, including background estimation, source detection,
+deblending and a wide array of source measurements, and finally
+outputs a FITS format catalog file.
+
+While Source Extractor is highly useful, the fact that it can only be
+used as an executable can limit its applicability or lead to awkward
+workflows. There is often a desire to have programmatic access to
+perform one or more of the above tasks on in-memory images as part of
+a larger custom analysis.
+
+**SEP makes the core algorithms of Source Extractor available as a
+library of stand-alone functions and classes.** These operate directly
+on in-memory arrays (no FITS files or configuration files).  The code
+is derived from the Source Extractor code base (written in C) and aims
+to produce results compatible with Source Extractor whenever possible.
+SEP consists of a C library with no dependencies outside the standard
+library, and a Python module that wraps the C library in a Pythonic
+API. The Python wrapper operates on NumPy arrays with NumPy as its
+only dependency. See below for language-specfic build and usage
+instructions.
 
 **Some features:**
 
@@ -17,22 +41,31 @@ directly from (and tested against) the Source Extractor code base.
 
 **Additional features not in Source Extractor:**
 
-- Optimized matched filter for variable noise in source extraction
-- circular annulus and elliptical annulus apertures
-- Local background subtraction in shape consistent with aperture
-- exact pixel overlap mode in all aperture photometry functions
-- ellipse masking
+- Optimized matched filter for variable noise in source extraction.
+- Circular annulus and elliptical annulus aperture photometry functions.
+- Local background subtraction in shape consistent with aperture in aperture
+  photometry functions.
+- Exact pixel overlap mode in all aperture photometry functions.
+- Masking of elliptical regions on images.
 
 
 Installation
 ------------
 
-SEP supports both Python 3 and Python 2 and requires only numpy.
-To install::
+SEP can be installed with `pip <https://pip.pypa.io>`_. After ensuring
+that numpy is installed, run ::
 
     pip install --no-deps sep
 
-**Development version / source code:** http://github.com/kbarbary/sep
+If you get an error about permissions, you are probably using your
+system Python. In this case, I recommend using `pip's "user install"
+<https://pip.pypa.io/en/latest/user_guide/#user-installs>`_ option to
+install sep into your user directory ::
+
+    pip install --no-deps --user sep
+
+Do **not** install sep or other third-party Python packages using ``sudo``
+unless you are fully aware of the risks.
 
 
 Usage Guide
@@ -52,6 +85,18 @@ Usage Guide
 For complete API documentation, see :doc:`reference`.
 
 
+Contributing
+------------
+
+Report a bug or documentation issue: http://github.com/kbarbary/sep/issues
+
+Development of SEP takes place on GitHub at
+http://github.com/kbarbary/sep.  Contributions of bug fixes,
+documentation improvements and minor feature additions are welcome via
+GitHub pull requests. For major features, it is best to open an issue
[…]
