### 2016-09-23 f3535f1
move matched filter docs to own page
