### 2016-09-23 55aeadf
fix bug in Background.rms()
