### 2016-09-23 0255850
remove duplicated code in background evaluation
