+++ src/background.c
@@ -758,9 +758,13 @@ float sep_bkg_pix(sep_bkg *bkg, int x, int y)
 
 /*****************************************************************************/
 
-int sep_bkg_line_flt(sep_bkg *bkg, int y, float *line)
+int bkg_line_flt_internal(sep_bkg *bkg, float *values, float *dvalues, int y,
+                          float *line)
 /* Interpolate background at line y (bicubic spline interpolation between
- * background map vertices) and save to line */
+ * background map vertices) and save to line.
+ * (values, dvalues) is either (bkg->back, bkg->dback) or
+ * (bkg->sigma, bkg->dsigma) depending on whether the background value or rms
+ * is being evaluated. */
 {
   int i,j,x,yl, nbx,nbxm1,nby, nx,width, ystep, changepoint, status;
   float	dx,dx0,dy,dy3, cdx,cdy,cdy3, temp, xstep;
@@ -795,9 +799,9 @@ int sep_bkg_line_flt(sep_bkg *bkg, int y, float *line)
       dy3 = (dy*dy*dy-dy);
       cdy3 = (cdy*cdy*cdy-cdy);
       ystep = nbx*yl;
-      blo = bkg->back + ystep;
+      blo = values + ystep;
       bhi = blo + nbx;
-      dblo = bkg->dback + ystep;
+      dblo = dvalues + ystep;
       dbhi = dblo + nbx;
       QMALLOC(nodebuf, float, nbx, status);  /* Interpolated background */
       nodep = node = nodebuf;
@@ -834,8 +838,8 @@ int sep_bkg_line_flt(sep_bkg *bkg, int y, float *line)
   else
     {
       /*-- No interpolation and no new 2nd derivatives needed along y */
-      node = bkg->back;
-      dnode = bkg->dback;
+      node = values;
+      dnode = dvalues;
     }
 
   /*-- Interpolation along x */
@@ -885,130 +889,20 @@ int sep_bkg_line_flt(sep_bkg *bkg, int y, float *line)
   return status;
 }
 
+int sep_bkg_line_flt(sep_bkg *bkg, int y, float *line) 
+/* Interpolate background at line y (bicubic spline interpolation between
+ * background map vertices) and save to line */
+{
+  return bkg_line_flt_internal(bkg, bkg->back, bkg->dback, y, line);
+}
 
 /*****************************************************************************/
 
 int sep_bkg_rmsline_flt(sep_bkg *bkg, int y, float *line)
-/* Bicubic-spline interpolation of the background noise along the current
- * scanline (y). NOTE: Most of the code is a copy of subbackline(), for
- * optimization reasons.
- */
+/* Interpolate background rms at line y (bicubic spline interpolation between
+ * background map vertices) and save to line */
 {
-  int i,j,x,yl, nbx,nbxm1,nby, nx, width, ystep, changepoint, status;
-  float	dx,dx0,dy,dy3, cdx,cdy,cdy3, temp, xstep;
-  float *nodebuf, *dnodebuf, *u;
-  float *node, *nodep, *dnode, *blo, *bhi, *dblo, *dbhi;
-  status = RETURN_OK;
-  nodebuf = node = NULL;
-  dnodebuf = dnode = NULL;
-  u = NULL;
-
-  width = bkg->w;
-  nbx = bkg->nx;
-  nbxm1 = nbx - 1;
-  nby = bkg->ny;
-  if (nby > 1)
[…]
