### 2016-09-23 f9b8a51
document bugfix in CHANGES.md
