### 2016-09-23 ff6e79c
try using external numpydoc
