+++ docs/conf.py
@@ -25,7 +25,7 @@ if not os.path.exists('api'):
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #sys.path.insert(0, os.path.abspath('.'))
-sys.path.append(os.path.abspath('ext'))
+#sys.path.append(os.path.abspath('ext'))
 
 # -- General configuration ------------------------------------------------
 
