### 2016-09-23 #58 @kbarbary
Thanks for the excellent feedback @danielskatz!

- **Installation**

  I updated the install instructions in both the README and documentation. Here is the part relevant to your problem:

  "If you get an error about permissions, you are probably using your system Python. In this case, I recommend using [pip's "user install"](https://pip.pypa.io/en/latest/user_guide/#user-installs) option to install sep into your user directory:

  ```
  pip install --no-deps --user sep
  ```

  Do **not** install sep or other third-party Python packages using `sudo` unless you are fully aware of the risks."

- **Contributing**

  I added a short paragraph on contributing to the README and docs. (It's basically just "use github.")

- **Statement of need in README**

  I copied a bunch of the paper content into the README and the docs to provide more context, as suggested.

- **License info in docs**

  I think this is OK as is. The docs are aimed at users, who only need to know that using sep (the Python library) is governed by LGPLv3. Developers reading the README might want to know that some files have more liberal licenses, so the README has a more detailed explanation of licenses.

- **Tutorial**

  I added a tutorial to the docs: http://sep.readthedocs.io/en/latest/tutorial.html
