### 2016-09-24 #58 @danielskatz
Thanks - this enables me to check another box and succeed in the install.

However, there are still some issues:

In the README "execute ./test.py" is a bit misleading.  Perhaps "execute python ./text.py" would be better.

And when I run that, I don't see any result.  No message is returned.  No files are created.  Perhaps this is because I don't have a FITS reader installed.  But even if so, this should say something.  And the docs should suggest that I install one for testing, I think.
