### 2016-09-24 #58 @kbarbary
The `py.test` executable is provided by the `pytest` package, which is required to run the tests. The README does mention this requirement ("requires the pytest package"), but it could be more prominent.
