### 2016-09-24 f19ca5b
make pytest requirement more prominent; mark FITS-skipped test
