### 2016-09-24 #58 @danielskatz
I have pytest installed (via pip). Perhaps I need to do something with my path?
