### 2016-09-24 #58 @danielskatz
ok, resolved - not sure if this should be part of the docs or not.

Now that I have this working, I get 20 successes and 1 failure:

tmp3:sep dsk$ ./test.py
platform darwin -- Python 2.7.10, pytest-3.0.2, py-1.4.31, pluggy-0.3.1
rootdir: /Users/dsk/Desktop/sep, inifile:
collected 21 items

test.py F....................

=================================== FAILURES ===================================
______________________________ test_vs_sextractor ______________________________

    @pytest.mark.skipif(NO_FITS, reason="no FITS reader")
    def test_vs_sextractor():
        """Test behavior of sep versus sextractor.

        Note: we turn deblending off for this test. This is because the
        deblending algorithm uses a random number generator. Since the sequence
        of random numbers is not the same between sextractor and sep or between
        different platforms, object member pixels (and even the number of objects)
        can differ when deblending is on.

        Deblending is turned off by setting DEBLEND_MINCONT=1.0 in the sextractor
        configuration file and by setting deblend_cont=1.0 in sep.extract().
        """

        data = np.copy(image_data)  # make an explicit copy so we can 'subfrom'
        bkg = sep.Background(data, bw=64, bh=64, fw=3, fh=3)

        # Test that SExtractor background is same as SEP:
        bkgarr = bkg.back(dtype=np.float32)
        assert_allclose(bkgarr, image_refback, rtol=1.e-5)

            # Test that SExtractor background rms is same as SEP:
        rmsarr = bkg.rms(dtype=np.float32)
>       assert_allclose(rmsarr, image_refrms, rtol=1.e-4)

test.py:90:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
/System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy/testing/utils.py:1181: in assert_allclose
    verbose=verbose, header=header)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

comparison = <function compare at 0x10e1b21b8>
x = array([[-1.90392494, -1.72829401, -1.56381047, ..., -1.22583032,
        -1.35....55419147, ..., -1.28184831,
        -1.41667461, -1.56063807]], dtype=float32)
y = array([[ 67.97720337,  67.97009277,  67.96318817, ...,  62.05981064,
         ...062103, ...,  65.2692337 ,
         65.27049255,  65.27170563]], dtype=float32)
err_msg = '', verbose = True
header = 'Not equal to tolerance rtol=0.0001, atol=0'

    def assert_array_compare(comparison, x, y, err_msg='', verbose=True,
                             header=''):
        from numpy.core import array, isnan, isinf, any, all, inf
        x = array(x, copy=False, subok=True)
        y = array(y, copy=False, subok=True)

        def isnumber(x):
            return x.dtype.char in '?bhilqpBHILQPefdgFDG'

        def chk_same_position(x_id, y_id, hasval='nan'):
            """Handling nan/inf: check that x and y have the nan/inf at the same
            locations."""
            try:
                assert_array_equal(x_id, y_id)
            except AssertionError:
                msg = build_err_msg([x, y],
                                    err_msg + '\nx and y %s location mismatch:' \
                                    % (hasval), verbose=verbose, header=header,
                                    names=('x', 'y'))
                raise AssertionError(msg)

        try:
            cond = (x.shape==() or y.shape==()) or x.shape == y.shape
            if not cond:
                msg = build_err_msg([x, y],
                                    err_msg
                                    + '\n(shapes %s, %s mismatch)' % (x.shape,
                                                                      y.shape),
                                    verbose=verbose, header=header,
                                    names=('x', 'y'))
                if not cond :
                    raise AssertionError(msg)

            if isnumber(x) and isnumber(y):
                x_isnan, y_isnan = isnan(x), isnan(y)
                x_isinf, y_isinf = isinf(x), isinf(y)

                # Validate that the special values are in the same place
                if any(x_isnan) or any(y_isnan):
                    chk_same_position(x_isnan, y_isnan, hasval='nan')
                if any(x_isinf) or any(y_isinf):
                    # Check +inf and -inf separately, since they are different
                    chk_same_position(x == +inf, y == +inf, hasval='+inf')
                    chk_same_position(x == -inf, y == -inf, hasval='-inf')

                # Combine all the special values
                x_id, y_id = x_isnan, y_isnan
                x_id |= x_isinf
                y_id |= y_isinf

                # Only do the comparison if actual values are left
                if all(x_id):
                    return

                if any(x_id):
                    val = comparison(x[~x_id], y[~y_id])
                else:
                    val = comparison(x, y)
            else:
                val = comparison(x, y)

            if isinstance(val, bool):
                cond = val
                reduced = [0]
            else:
                reduced = val.ravel()
                cond = reduced.all()
                reduced = reduced.tolist()
            if not cond:
                match = 100-100.0*reduced.count(1)/len(reduced)
                msg = build_err_msg([x, y],
                                    err_msg
                                    + '\n(mismatch %s%%)' % (match,),
                                    verbose=verbose, header=header,
                                    names=('x', 'y'))
                if not cond :
>                   raise AssertionError(msg)
E                   AssertionError:
E                   Not equal to tolerance rtol=0.0001, atol=0
E
E                   (mismatch 100.0%)
E                    x: array([[-1.90392494, -1.72829401, -1.56381047, ..., -1.22583032,
E                           -1.3547647 , -1.49243712],
E                          [-1.90392673, -1.7282958 , -1.56381226, ..., -1.22646201,...
E                    y: array([[ 67.97720337,  67.97009277,  67.96318817, ...,  62.05981064,
E                            62.02336884,  61.98696136],
E                          [ 67.93690491,  67.93002319,  67.9233551 , ...,  62.09538269,...

/System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy/testing/utils.py:644: AssertionError
===================== 1 failed, 20 passed in 1.37 seconds ======================
