### 2016-09-24 #58 @danielskatz
Is it worth updating the paper for the new version?

I think that would be best, but if not, I will approve this one.
