### 2016-09-24 #58 @kbarbary
I agree. I'll do the release and post here after the paper is updated (including a new zenodo DOI).
