### 2016-09-30 8c7c28d
enter date for v1.0 release
