### 2016-10-04 #58 @kbarbary
SEP v1.0.0 DOI: http://dx.doi.org/10.5281/zenodo.159035
