### 2016-10-04 #58 @arfon
@danielskatz by the looks of the checklist I think we're good to accept here. Can you confirm?
