### 2016-10-05 #58 @arfon
> 1 - in the paper, I see "Software Archive: PENDING" This should be updated to http://dx.doi.org/10.5281/zenodo.159035
> 2 - in the metadata at the top of this issue, I see "Version: v0.6.0", which should now be 1.0.0

👍 I can do these two things.

> Otherwise, I'm happy to accept this.

🎉
