### 2016-10-05 #58 @kbarbary
Thanks @arfon and thanks @danielskatz for the review; it improved the package!
